+++ src/main/java/com/kurkus/kusinsa/controller/product/ProductStatusController.java
@@ -1,0 +1,4 @@
+package com.kurkus.kusinsa.controller.product;
+
+public class ProductStockController {
+}
