+++ src/main/java/com/kurkus/kusinsa/service/product/ProductStatusService.java
@@ -1,0 +1,2 @@
+package com.kurkus.kusinsa.service.product;public class ProductStatusService {
+}
